+++ src/lib/utils.ts
@@ -1,6 +1,8 @@
 import { clsx, type ClassValue } from 'clsx';
 import dayjs, { type Dayjs } from 'dayjs';
 import { twMerge } from 'tailwind-merge';
+import { type z } from 'zod';
+import { type step4SingleSchema } from '~/core/dashboard/calendar/new-event/components/Step4Single';
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
@@ -57,6 +59,42 @@ export function dateToTimeString(date: Date) {
 export function timeStringToMinutes(str: string) {
   const [hours, minutes] = str.split(':').map(Number) as [number, number];
   return hours * 60 + minutes;
+}
+
+export function getSessionTimespan({
+  session,
+}: {
+  session: z.infer<typeof step4SingleSchema>['sessions'][number];
+}) {
+  const start = timeStringToMinutes(session.start);
+  const end = 'end' in session ? timeStringToMinutes(session.end) : undefined;
+
+  console.log('session date', dayjs(session.date).format('YYYY/MM/DD HH:mm'));
+
+  const baseDate = dayjs(session.date)
+    .set('minutes', 0)
+    .set('hours', 0)
+    .set('seconds', 0);
+
+  console.log('base date', baseDate.format('YYYY/MM/DD HH:mm'));
+
+  const startDate = baseDate.add(start, 'minutes');
+  const endDate = end ? baseDate.add(end, 'minutes') : undefined;
+
+  console.log('startDate', startDate.format('YYYY/MM/DD HH:mm'));
+  console.log('endDate', endDate?.format('YYYY/MM/DD HH:mm'));
+
+  if ('endsNextDay' in session && session.endsNextDay && endDate) {
+    return {
+      start: startDate.toDate(),
+      end: endDate.add(24, 'hours').toDate(),
+    };
+  }
+
+  return {
+    start: startDate.toDate(),
+    end: endDate?.toDate(),
+  };
 }
 
 export function getCalendarRowStyles({
