+++ src/server/api/routers/event.ts
@@ -1,12 +1,11 @@
 import { createTRPCRouter, protectedProcedure } from '../trpc';
 import { step2SingleSchema } from '~/core/dashboard/calendar/new-event/components/Step2Single';
 import { step3SingleSchema } from '~/core/dashboard/calendar/new-event/components/Step3Single';
-import { step4SingleSchema } from '~/core/dashboard/calendar/new-event/components/Step4Single';
 import { timeStringToDate, type ReplaceAll } from '~/lib/utils';
 import { z } from 'zod';
-import { encryptString, getSessionTimespan } from '../utils/utils';
+import { encryptString } from '../utils/utils';
 import { games } from '~/lib/constants';
-import dayjs from 'dayjs';
+import { sessionSchema } from '~/core/dashboard/calendar/new-event/components/SessionForm';
 
 export const eventRouter = createTRPCRouter({
   createOrEdit: protectedProcedure
@@ -17,7 +16,16 @@ export const eventRouter = createTRPCRouter({
             eventType: z.literal('single'),
             stepTwo: step2SingleSchema,
             stepThree: step3SingleSchema,
-            stepFour: step4SingleSchema,
+            stepFour: z.object({
+              sessions: z.array(
+                sessionSchema.and(
+                  z.object({
+                    startDate: z.date(),
+                    endDate: z.date().optional(),
+                  })
+                )
+              ),
+            }),
           }),
           z.object({
             eventType: z.literal('championship'),
@@ -39,11 +47,6 @@ export const eventRouter = createTRPCRouter({
           stepThree: { rosterId },
           stepFour: { sessions },
         } = input;
-
-        console.log(
-          'session dates',
-          sessions.map(s => dayjs(s.date).format('YYYY/MM/DD HH:mm'))
-        );
 
         const event = await ctx.prisma.event.upsert({
           where: { id: eventId ?? '' },
@@ -97,15 +100,10 @@ export const eventRouter = createTRPCRouter({
               ? [session.driverId]
               : [];
 
-          console.log(
-            dayjs(getSessionTimespan({ session }).start).format(
-              'YYYY/MM/DD HH:mm'
-            )
-          );
-
           await ctx.prisma.eventSession.create({
             data: {
-              ...getSessionTimespan({ session }),
+              start: session.startDate,
+              end: session.endDate,
               type: session.type,
               event: { connect: { id: event.id } },
               drivers:
